refactor: create a seperate batchItemTable component
related to #41688
