+++ operate/client/src/App/BatchOperation/index.test.tsx
@@ -17,12 +17,10 @@ import {
 import {MemoryRouter, Route, Routes} from 'react-router-dom';
 import {QueryClientProvider} from '@tanstack/react-query';
 import {Paths} from 'modules/Routes';
-import type {
-  BatchOperationItem,
-  BatchOperation as BatchOperationType,
-} from '@camunda/camunda-api-zod-schemas/8.8';
+import type {BatchOperation as BatchOperationType} from '@camunda/camunda-api-zod-schemas/8.8';
 import {mockQueryBatchOperationItems} from 'modules/mocks/api/v2/batchOperations/queryBatchOperationItems';
 import {mockGetBatchOperation} from 'modules/mocks/api/v2/batchOperations/getBatchOperation';
+import {notificationsStore} from 'modules/stores/notifications';
 
 const operation: BatchOperationType = {
   batchOperationKey: 'migrate-operation-123',
@@ -42,29 +40,20 @@ const Wrapper: React.FC<{children?: React.ReactNode}> = ({children}) => (
     >
       <Routes>
         <Route path={Paths.batchOperation()} element={children} />
+        <Route
+          path={Paths.batchOperations()}
+          element={<div data-testid="batch-operations" />}
+        />
       </Routes>
     </MemoryRouter>
   </QueryClientProvider>
 );
 
-const createMockBatchOperationItems = (
-  count: number,
-  options?: {
-    withErrors?: boolean;
-    state?: BatchOperationItem['state'];
+vi.mock('modules/stores/notifications', () => ({
+  notificationsStore: {
+    displayNotification: vi.fn(() => () => {}),
   },
-): BatchOperationItem[] => {
-  return Array.from({length: count}, (_, i) => ({
-    batchOperationKey: `migrate-operation-${i + 1}`,
-    itemKey: `item-${i + 1}`,
-    processInstanceKey: `2251799813685${250 + i}`,
-    state: options?.state ?? 'COMPLETED',
-    processedDate: '2023-11-22T09:03:29.564+0100',
-    errorMessage: options?.withErrors
-      ? `Error processing item ${i + 1}`
-      : undefined,
-  }));
-};
+}));
 
 describe('<BatchOperation />', () => {
   beforeEach(() => {
@@ -76,25 +65,23 @@ describe('<BatchOperation />', () => {
     });
   });
 
-  it('should render skeleton state when loading', async () => {
+  it('should render items count summary', async () => {
     render(<BatchOperation />, {
       wrapper: Wrapper,
     });
 
-    expect(screen.getByTestId('data-table-skeleton')).toBeInTheDocument();
-
     await waitForElementToBeRemoved(() =>
-      screen.queryAllByTestId('data-table-skeleton'),
+      screen.queryAllByTestId('text-skeleton'),
     );
 
-    expect(screen.queryByTestId('data-table-skeleton')).not.toBeInTheDocument();
+    expect(screen.getByText(/2 successful/)).toBeInTheDocument();
   });
 
   it('should render page title with formatted operation type', async () => {
     render(<BatchOperation />, {wrapper: Wrapper});
 
     await waitForElementToBeRemoved(() =>
-      screen.queryAllByTestId('data-table-skeleton'),
+      screen.queryAllByTestId('text-skeleton'),
     );
 
     expect(
@@ -106,7 +93,7 @@ describe('<BatchOperation />', () => {
     render(<BatchOperation />, {wrapper: Wrapper});
 
     await waitForElementToBeRemoved(() =>
-      screen.queryAllByTestId('data-table-skeleton'),
+      screen.queryAllByTestId('text-skeleton'),
     );
 
     expect(screen.getByText('State')).toBeInTheDocument();
@@ -119,7 +106,7 @@ describe('<BatchOperation />', () => {
     render(<BatchOperation />, {wrapper: Wrapper});
 
     await waitForElementToBeRemoved(() =>
-      screen.queryAllByTestId('data-table-skeleton'),
+      screen.queryAllByTestId('text-skeleton'),
     );
 
     expect(screen.getByText('Completed')).toBeInTheDocument();
@@ -129,92 +116,11 @@ describe('<BatchOperation />', () => {
     render(<BatchOperation />, {wrapper: Wrapper});
 
     await waitForElementToBeRemoved(() =>
-      screen.queryAllByTestId('data-table-skeleton'),
+      screen.queryAllByTestId('text-skeleton'),
     );
 
     const tiles = screen.getAllByText(/2021|2023/);
     expect(tiles.length).toBeGreaterThan(0);
-  });
-
-  it('should render items count summary', async () => {
-    const items = createMockBatchOperationItems(2);
-    mockQueryBatchOperationItems().withSuccess({
-      items,
-      page: {totalItems: 2},
-    });
-    render(<BatchOperation />, {wrapper: Wrapper});
-
-    await waitForElementToBeRemoved(() =>
-      screen.queryAllByTestId('data-table-skeleton'),
-    );
-
-    expect(screen.getByText(/2 successful/)).toBeInTheDocument();
-  });
-
-  it('should render empty state when no items exist', async () => {
-    render(<BatchOperation />, {wrapper: Wrapper});
-
-    await waitForElementToBeRemoved(() =>
-      screen.queryAllByTestId('data-table-skeleton'),
-    );
-
-    expect(screen.getByText('0 Items')).toBeInTheDocument();
-    expect(screen.getByText('No items found')).toBeInTheDocument();
-  });
-
-  it('should render table with batch operation items', async () => {
-    const items = createMockBatchOperationItems(3);
-    mockQueryBatchOperationItems().withSuccess({
-      items,
-      page: {totalItems: 3},
-    });
-
-    render(<BatchOperation />, {wrapper: Wrapper});
-
-    await waitForElementToBeRemoved(() =>
-      screen.queryAllByTestId('data-table-skeleton'),
-    );
-
-    expect(screen.getByText('3 Items')).toBeInTheDocument();
-    expect(screen.getByText('2251799813685250')).toBeInTheDocument();
-    expect(screen.getByText('2251799813685251')).toBeInTheDocument();
-    expect(screen.getByText('2251799813685252')).toBeInTheDocument();
-  });
-
-  it('should render items with failed state', async () => {
-    const items = createMockBatchOperationItems(2, {state: 'FAILED'});
-    mockQueryBatchOperationItems().withSuccess({
-      items,
-      page: {totalItems: 2},
-    });
-
-    render(<BatchOperation />, {wrapper: Wrapper});
-
-    await waitForElementToBeRemoved(() =>
-      screen.queryAllByTestId('data-table-skeleton'),
-    );
-
-    const failedStates = screen.getAllByText('Failed');
-    expect(failedStates.length).toBeGreaterThan(0);
-  });
-
-  it('should render error messages for failed items', async () => {
-    const items = createMockBatchOperationItems(1, {
-      withErrors: true,
-      state: 'FAILED',
-    });
-    mockQueryBatchOperationItems().withSuccess({
-      items,
-      page: {totalItems: 1},
-    });
-
-    render(<BatchOperation />, {wrapper: Wrapper});
-
-    await waitForElementToBeRemoved(() =>
-      screen.queryAllByTestId('data-table-skeleton'),
-    );
-    expect(screen.getByText('Failure reason:')).toBeInTheDocument();
-    expect(screen.getByText('Error processing item 1')).toBeInTheDocument();
   });
 
   it('should show error notification when batch operation fails to load', async () => {
@@ -235,7 +141,7 @@ describe('<BatchOperation />', () => {
     render(<BatchOperation />, {wrapper: Wrapper});
 
     await waitForElementToBeRemoved(() =>
-      screen.queryAllByTestId('data-table-skeleton'),
+      screen.queryAllByTestId('text-skeleton'),
     );
 
     expect(
@@ -243,5 +149,21 @@ describe('<BatchOperation />', () => {
         '403 - You do not have permission to view this information',
       ),
     ).toBeInTheDocument();
+  });
+
+  it('should redirect to batch operations page and display notification if batch operation is not found (404)', async () => {
+    mockGetBatchOperation().withServerError(404);
+
+    render(<BatchOperation />, {wrapper: Wrapper});
+
+    await waitFor(() => {
+      expect(screen.getByTestId('batch-operations')).toBeInTheDocument();
+    });
+
+    expect(notificationsStore.displayNotification).toHaveBeenCalledWith({
+      kind: 'error',
+      title: 'batch operation migrate-operation-123 could not be found',
+      isDismissable: true,
+    });
   });
 });
